Add files via upload
Wrote a reroll function that allows the player to choose one die to reroll, also implements a function for the CPU opponent to reroll one die as well. Additionally, added a welcome message for the player at start up that lists the rules.  Began work on a bank account class for the player and opponent although I haven't decided how it will work yet.

Daniel Sipe
danielbsipe@gmail.com